\docType{package}
\name{messina-package}
\alias{messina}
\alias{messina-package}
\title{TODO}
\description{
  TODO
}
\details{
  \tabular{11}{ Package: \tab messina\cr Type: \tab
  Package\cr Version: \tab 0.99.0\cr Date: \tab TODO\cr
  License: \tab EPL (>= 1.0)\cr LazyLoad: \tab yes\cr }

  Using the Fermat primality test, pseudoprime checks for
  primes probabilistically; the test is fooled every time
  by Carmichael numbers.

  \code{\link{is.pseudoprime}} checks a number \code{n} for
  pseudoprimality, applying Fermat?s test \code{times}
  times.
}
\author{
  Mark Pinese \email{m.pinese@garvan.org.au}
}
\references{
  \url{TODO}
}
\seealso{
  \code{\link{TODO}}
}
\keyword{package}

