% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,MessinaResult-method}
\alias{show,MessinaFits-method}
\alias{show,MessinaParameters-method}
\alias{show,MessinaResult-method}
\title{Generic show methods for Messina objects.}
\usage{
\S4method{show}{MessinaResult}(object)

\S4method{show}{MessinaParameters}(object)

\S4method{show}{MessinaFits}(object)
}
\description{
Generic show methods for Messina objects.
}
\details{
For details of the objects and their generation, see the
relevant class documentation, and entries for the main
functions \code{\link{messina}}, \code{\link{messinaDE}},
and \code{\link{messinaSurv}},
}
\author{
Mark Pinese \email{m.pinese@garvan.org.au}
}
\seealso{
\code{\link{MessinaResult-class}}

\code{\link{MessinaParameters-class}}

\code{\link{MessinaFits-class}}

\code{\link{messina}}

\code{\link{messinaDE}}

\code{\link{messinaSurv}}
}

