\name{summary.MessinaResult}
\alias{summary.MessinaResult}
\title{Print a summary of the results of a Messina fit.}
\usage{
  summary.MessinaResult(result, ...)
}
\arguments{
  \item{result}{an object of class MessinaResult containing
  the result of a messina fit, as returned by the function
  messina.}
}
\description{
  TODO
}
\author{
  Mark Pinese \email{m.pinese@garvan.org.au}
}
\references{
  TODO
}
\seealso{
  \code{\link{MessinaResult-class}}

  \code{\link{messina}}

  \code{\link{plot.MessinaResult}}
}

