\name{toptable.MessinaResult}
\alias{toptable.MessinaResult}
\title{Print the results of a Messina fit, ordered by decreasing margin size.}
\usage{
  toptable.MessinaResult(result, ...)
}
\arguments{
  \item{result}{an object of class MessinaResult containing
  the result of a messina fit, as returned by the function
  messina.}

  \item{maxn}{the maximum number of top features to
  display.  If missing, all features are printed.}

  \item{minmar}{only features with a fit margin of at least
  minmar will be printed.}
}
\description{
  TODO
}
\author{
  Mark Pinese \email{m.pinese@garvan.org.au}
}
\references{
  TODO
}
\seealso{
  \code{\link{MessinaResult-class}}

  \code{\link{messina}}

  \code{\link{plot.MessinaResult}}
}

