% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/messina-package.R
\docType{package}
\name{messina-package}
\alias{messina-package}
\title{The Messina package for classification and outlier differential expression.}
\description{
Single-gene classifiers and outlier-resistant detection of
differential expression for two-group and survival
problems.
}
\details{
Messina is a collection of algorithms for constructing
optimally robust single-gene classifiers, and for
identifying differential expression in the presence of
outliers or unknown sample subgroups.  The methods have
application in identifying lead features to develop into
clinical tests (both diagnostic and prognostic), and in
identifying differential expression when a fraction of
samples show unusual patterns of expression.
}
\author{
Mark Pinese \email{m.pinese@garvan.org.au}
}
\references{
Pinese M, Scarlett CJ, Kench JG, et al. (2009) Messina: A
Novel Analysis Tool to Identify Biologically Relevant
Molecules in Disease.  PLoS ONE 4(4): e5337.
\url{doi:10.1371/journal.pone.0005337}
}
\seealso{
\code{\link{messina}}

\code{\link{messinaDE}}

\code{\link{messinaSurv}}
}
\keyword{package}

